fix: API 태그와 엔드포인트 정리
