+++ src/main/java/com/waggle/domain/user/service/UserServiceImpl.java
@@ -86,129 +86,9 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Set<Project> getCurrentUserProjects() {
-        User user = getCurrentUser();
-        return user.getProjectMembers().stream()
-                .map(ProjectMember::getProject)
-                .sorted(Comparator.comparing(Project::getCreatedAt).reversed())
-                .collect(Collectors.toCollection(LinkedHashSet::new));
-    }
-
-    @Override
-    public boolean toggleCurrentUserBookmark(String projectId) {
-        User user = getCurrentUser();
-        Project project = projectRepository.findById(UUID.fromString(projectId))
-                .orElseThrow(() -> new EmptyResultDataAccessException(1));
-        boolean isBookmarked;
-
-        if (user.getProjectBookmarks().stream().anyMatch(projectBookmark -> projectBookmark.getProject().getId().equals(project.getId()))) {
-            user.getProjectBookmarks().removeIf(projectBookmark -> projectBookmark.getProject().getId().equals(project.getId()));
-            project.setBookmarkCnt(project.getBookmarkCnt() - 1);
-            isBookmarked = false;
-        } else {
-            user.getProjectBookmarks().add(ProjectBookmark.builder()
-                    .project(project)
-                    .user(user)
-                    .build());
-            project.setBookmarkCnt(project.getBookmarkCnt() + 1);
-            isBookmarked = true;
-        }
-
-        userRepository.save(user);
-        projectRepository.save(project);
-
-        return isBookmarked;
-    }
-
-    @Override
-    public Set<Project> getCurrentUserBookmarkProjects() {
-        User user = getCurrentUser();
-        return user.getProjectBookmarks().stream()
-                .map(ProjectBookmark::getProject)
-                .sorted(Comparator.comparing(Project::getCreatedAt).reversed())
-                .collect(Collectors.toCollection(LinkedHashSet::new));
-    }
-
-    @Override
     public User getUserByUserId(String userId) {
         return userRepository.findByUserId(UUID.fromString(userId))
                 .orElseThrow(() -> new EmptyResultDataAccessException(1));
-    }
-
-    @Override
-    public Set<Project> getUserProjects(String userId) {
-        User user = getUserByUserId(userId);
-        return user.getProjectMembers().stream()
-                .map(ProjectMember::getProject)
-                .sorted(Comparator.comparing(Project::getCreatedAt).reversed())
-                .collect(Collectors.toCollection(LinkedHashSet::new));
-    }
-
-    @Override
-    @Transactional
-    public void deleteUserProject(String projectId) {
-        User user = getCurrentUser();
-        Project project = projectRepository.findById(UUID.fromString(projectId))
-                .orElseThrow(() -> new EmptyResultDataAccessException(1));
-
-        if (!project.getProjectMembers().stream().anyMatch(projectMember -> projectMember.getUser().getId().equals(user.getId()))) {
-            throw new ProjectException(ApiStatus._NOT_JOINED_PROJECT);
-        }
-
-        project.getProjectMembers().removeIf(projectMember -> projectMember.getUser().getId().equals(user.getId()));
-        projectRepository.save(project);
-    }
-
-    @Override
-    public Set<Project> getUserBookmarkProjects(String userId) {
-        User user = getUserByUserId(userId);
-        return user.getProjectBookmarks().stream()
-                .map(ProjectBookmark::getProject)
-                .sorted(Comparator.comparing(Project::getCreatedAt).reversed())
-                .collect(Collectors.toCollection(LinkedHashSet::new));
-    }
-
-    @Override
-    public Project applyProject(String projectId) {
-        User user = getCurrentUser();
-        Project project = projectRepository.findById(UUID.fromString(projectId))
-                .orElseThrow(() -> new EmptyResultDataAccessException(1));
-
-        if (project.getProjectMembers().stream().anyMatch(projectMember -> projectMember.getUser().getId().equals(user.getId()))) {
-            throw new ProjectException(ApiStatus._ALREADY_JOINED_PROJECT);
-        }
-
-        if (project.getProjectApplicants().stream().anyMatch(projectMember -> projectMember.getUser().getId().equals(user.getId()))) {
-            throw new ProjectException(ApiStatus._ALREADY_APPLIED_PROJECT);
-        }
-
-        project.getProjectApplicants().add(ProjectApplicant.builder()
-                .project(project)
-                .user(user)
-                .build());
-
-        projectRepository.save(project);
-        return project;
-    }
-
-    @Override
-    @Transactional
-    public void cancelApplyProject(String projectId) {
-        User user = getCurrentUser();
-        Project project = projectRepository.findById(UUID.fromString(projectId))
-                .orElseThrow(() -> new EmptyResultDataAccessException(1));
-
-        project.getProjectApplicants().removeIf(projectApplicant -> projectApplicant.getUser().getId().equals(user.getId()));
-        projectRepository.save(project);
-    }
-
-    @Override
-    public Set<Project> getAppliedProjects() {
-        User user = getCurrentUser();
-        return user.getProjectApplicants().stream()
-                .sorted(Comparator.comparing(ProjectApplicant::getAppliedAt).reversed())
-                .map(ProjectApplicant::getProject)
-                .collect(Collectors.toCollection(LinkedHashSet::new));
     }
 
     private Set<UserJob> getUserJobs(UserInputDto userInputDto, User user) {
